pyproject.toml
[tool.poetry]
name = "openvault"
version = "0.1.0"
description = ""
authors = ["Dessera <lord.changed@foxmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
tortoise-orm = {extras = ["asyncmy"], version = "^0.21.6"}
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = {extras = ["standard"], version = "^0.31.1"}
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.poetry.scripts]
openvault = "openvault.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
